remove Constants.js
