+++ frontend/src/Metamaps/Selected.js
@@ -1,0 +1,11 @@
+const Selected = {
+  reset: function () {
+    var self = Metamaps.Selected
+    self.Nodes = []
+    self.Edges = []
+  },
+  Nodes: [],
+  Edges: []
+}
+
+export default Selected
