+++ frontend/src/Metamaps/Settings.js
@@ -1,0 +1,21 @@
+const Settings = {
+  embed: false, // indicates that the app is on a page that is optimized for embedding in iFrames on other web pages
+  sandbox: false, // puts the app into a mode (when true) where it only creates data locally, and isn't writing it to the database
+  colors: {
+    background: '#344A58',
+    synapses: {
+      normal: '#888888',
+      hover: '#888888',
+      selected: '#FFFFFF'
+    },
+    topics: {
+      selected: '#FFFFFF'
+    },
+    labels: {
+      background: '#18202E',
+      text: '#DDD'
+    }
+  },
+}
+
+export default Settings
